Add slideshow support

+++ depressed_tiktok_bot/main.py
@@ -2,6 +2,9 @@ import asyncio
 import json
 import logging
 import os
+import io
+import re
+import base64
 import sys
 from os import getenv
 from urllib.parse import urlparse
@@ -11,7 +14,7 @@ from aiogram.enums import ParseMode
 from aiogram.filters import CommandStart
 from aiogram.types import Message
 from aiogram.utils.markdown import hbold
-from aiogram.types import FSInputFile
+from aiogram.types import FSInputFile, URLInputFile, InputMediaPhoto, InputMediaAudio
 
 from yt_dlp import YoutubeDL
 
@@ -23,18 +26,50 @@ TOKEN = getenv("BOT_TOKEN")
 # All handlers should be attached to the Router (or Dispatcher)
 dp = Dispatcher()
 
+def chunks(lst, n):
+    """Yield successive n-sized chunks from lst."""
+    for i in range(0, len(lst), n):
+        yield lst[i:i + n]
 
-@dp.message(CommandStart())
-async def command_start_handler(message: Message) -> None:
-    """
-    This handler receives messages with `/start` command
-    """
-    # Most event objects have aliases for API methods that can be called in events' context
-    # For example if you want to answer to incoming message you can use `message.answer(...)` alias
-    # and the target chat will be passed to :ref:`aiogram.methods.send_message.SendMessage`
-    # method automatically or call API method directly via
-    # Bot instance: `bot.send_message(chat_id=message.chat.id, ...)`
-    await message.answer(f"Hello, {hbold(message.from_user.full_name)}!")
+def check_if_slideshow(url):
+    with YoutubeDL() as ydl:
+        result = ydl.extract_info(url, download=False)
+
+        if result["formats"][0]["ext"] == ".mp3" or \
+            result["formats"][0]["url"] == ".m4a" or \
+            (not result["formats"][0]["width"] and not result["formats"][0]["height"]):
+            # this is most likely a slideshow
+            return True
+        
+    return False
+
+# Taken from https://github.com/dylanpdx/vxtiktok/blob/main/vxtiktok.py
+def get_slideshow_from_post_URL(url): # thsi function assumes the url is a slideshow
+    with YoutubeDL(params={"dump_intermediate_pages":True}) as ydl:
+        f = io.StringIO()
+        ydl._out_files.screen = f # this is a hack to get the output of ydl to a variable
+        result = ydl.extract_info(url, download=False)
+        s=f.getvalue()
+        # find the first line that's valid base64 data
+        data=None
+        for line in s.splitlines():
+            if re.match(r"^[A-Za-z0-9+/=]+$", line):
+                data= json.loads(base64.b64decode(line).decode())
+        thisPost = data["aweme_list"][0]
+        postMusicURL = thisPost["music"]["play_url"]["uri"]
+        postImages = thisPost["image_post_info"]["images"]
+        imageUrls=[]
+        for image in postImages:
+            for url in image["display_image"]["url_list"]:
+                imageUrls.append(url)
+                break
+
+        finalData = {
+            "musicURL": postMusicURL,
+            "imageURLs": imageUrls
+        }
+
+        return finalData
 
 async def download_and_reply(message: types.Message) -> None:
     parsed_url = urlparse(message.text)
@@ -42,15 +77,26 @@ async def download_and_reply(message: types.Message) -> None:
     ydl_opts = {
         'outtmpl': f'{tt_slug}.%(ext)s'
     }
-    with YoutubeDL(ydl_opts) as ydl:
-        def download_video():
-            ydl.download([message.text])
-        await asyncio.get_event_loop().run_in_executor(None, download_video)
-        # ℹ️ ydl.sanitize_info makes the info json-serializable
+    if not check_if_slideshow(message.text):
+        with YoutubeDL(ydl_opts) as ydl:
+            def download_video():
+                ydl.download([message.text])
+            await asyncio.get_event_loop().run_in_executor(None, download_video)
+            tg_file = FSInputFile(f'{tt_slug}.mp4')
+            await message.reply_video(tg_file)
+            os.remove(f'{tt_slug}.mp4')
+    else:
+        loop = asyncio.get_event_loop()
+        slideshow_data = await loop.run_in_executor(None, get_slideshow_from_post_URL, message.text)
+        tg_images = [InputMediaPhoto(media=URLInputFile(x)) for x in slideshow_data['imageURLs']]
+        tg_music = URLInputFile(slideshow_data['musicURL'])
 
-    tg_file = FSInputFile(f'{tt_slug}.mp4')
-    await message.reply_video(tg_file)
-    os.remove(f'{tt_slug}.mp4')
+        tg_images_chunked = list(chunks(tg_images, 10))
+        await message.reply_media_group(tg_images_chunked[0])
+        for chunk in tg_images_chunked:
+            await message.answer_media_group(chunk)
+        await message.answer_audio(tg_music)
+
 
 
 @dp.message()
@@ -68,8 +114,7 @@ async def echo_handler(message: types.Message) -> None:
             
             await download_and_reply(message)
     except TypeError:
-        # But not all the types is supported to be copied so need to handle it
-        await message.answer("Nice try!")
+        raise
 
 
 async def main() -> None:
